refactor(gallery): decompose gallery page to components
Placeholder, FluidGridLayout, Card
