+++ src/components/Gallery/Placeholder/Placeholder.module.scss
@@ -1,0 +1,28 @@
+.placeholder {
+  position: relative;
+  width: 100%;
+  height: calc(100% - 70px);
+  &_image {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+    vertical-align: middle;
+    opacity: 0.5;
+  }
+  &_title {
+    position: absolute;
+    top: 40%;
+    left: 50%;
+    transform: translate(-50%, -40%);
+    text-shadow: 1px 1px 7px #bdbdbd;
+    h1 {
+      color: white;
+      font-size: 5rem;
+      margin-bottom: 20px;
+    }
+    p {
+      color: white;
+      font-size: 2rem;
+    }
+  }
+}
