+++ src/components/Header/index.jsx
@@ -1,17 +1,15 @@
 import React from 'react';
 import { NavLink } from 'react-router-dom';
 
-import Search from './Search';
-
+import Search from '@/components/ui/Search';
 import styles from './Header.module.scss';
-
 import WeatherLogo from '@/assets/weather-logo.png';
 
 const Header = () => {
   return (
     <div className={styles.header}>
       <NavLink to='/' className={styles.header_logo}>
-        <img src={WeatherLogo} alt='' />
+        <img src={WeatherLogo} alt='logo' />
         Start of the day app
       </NavLink>
       <div className={styles.search_input}>
