--- src/pages/Collection.jsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Collection = () => {
-  return <div>Collection</div>;
-};
-
-export default Collection;
