+++ src/pages/Gallery/Gallery.module.scss
@@ -1,81 +1,8 @@
-.placeholder {
-  position: relative;
-  width: 100%;
-  height: calc(100% - 70px);
-  &_image {
-    width: 100%;
-    height: 100%;
-    object-fit: cover;
-    vertical-align: middle;
-    opacity: 0.5;
-  }
-  &_title {
-    position: absolute;
-    top: 40%;
-    left: 50%;
-    transform: translate(-50%, -40%);
-    text-shadow: 1px 1px 7px #bdbdbd;
-    h1 {
-      color: white;
-      font-size: 5rem;
-      margin-bottom: 20px;
-    }
-    p {
-      color: white;
-      font-size: 2rem;
-    }
-  }
-}
-
 .photos_container {
   max-width: 1500px;
   padding: 20px;
   margin: 0 auto;
   .collection_title {
     margin-bottom: 20px;
-  }
-  .photos {
-    column-width: 300px;
-    column-gap: 20px;
-    .photo_card {
-      width: 100%;
-      margin-bottom: 20px;
-      background-color: #f6f6f6;
-      border-radius: 10px;
-      break-inside: avoid;
-      cursor: pointer;
-      img {
-        width: 100%;
-        border-radius: 10px;
-        transition: 0.25s;
-        &:hover {
-          transform: scale(0.95);
-        }
-      }
-      .description {
-        padding: 15px 10px;
-        .name {
-          margin-bottom: 40px;
-          display: -webkit-box;
-          overflow: hidden;
-          -webkit-line-clamp: 3;
-          -webkit-box-orient: vertical;
-        }
-        .author {
-          font-style: italic;
-        }
-      }
-    }
-  }
-}
-
-.not_found_container {
-  width: 100%;
-  height: calc(100% - 70px);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  img {
-    max-width: 400px;
   }
 }
